+++ src/main/java/com/example/book_your_seat/queue/service/facade/QueueFacade.java
@@ -9,33 +9,29 @@ import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class QueueServiceImpl implements QueueService {
+public class QueueFacade {
     private final QueueCommandService queueCommandService;
     private final QueueQueryService queueQueryService;
 
-    @Override
+
     public TokenResponse issueTokenAndEnqueue(Long userId) {
         return new TokenResponse(queueCommandService.issueTokenAndEnqueue(userId));
     }
 
-    @Override
+
     public void dequeueWaitingQueue(Long userId, String token) {
         queueCommandService.removeTokenInWaitingQueue(userId, token);
     }
 
-    @Override
+
     public void dequeueProcessingQueue(Long userId, String token) {
         queueCommandService.removeTokenInProcessingQueue(userId, token);
     }
 
 
-    @Override
+
     public QueueResponse findQueueStatus(Long userId, String token) {
         return queueQueryService.findQueueStatus(userId, token);
     }
 
-    @Override
-    public void updateWaitingToProcessing() {
-        queueCommandService.updateWaitingToProcessing();
-    }
 }
